# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recetas DWES",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "operationId": "7edae2cdbaafc10dff8c9c4198b320a2",
                "responses": {
                    "200": {
                        "description": "Datos del usuario obtenidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Refrescar token de autenticación",
                "operationId": "e7ab7446393758bb71fd1946a349e500",
                "responses": {
                    "200": {
                        "description": "Token refrescado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recetas/{receta}/comentarios": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Listar comentarios de una receta",
                "operationId": "105ea8e0d855c888f15f9c80d1c406a0",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Crear un comentario",
                "operationId": "2317de860f1d4d22151aa7fb83e87502",
                "responses": {
                    "201": {
                        "description": "Comentario creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comentarios/{comentario}": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Ver un comentario específico",
                "operationId": "d6bb9cf15cdb0f625d810ba43fa63911",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Actualizar un comentario",
                "operationId": "8865fc4a424a9f2bcdf78fdf6a057d34",
                "responses": {
                    "200": {
                        "description": "Comentario actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Eliminar un comentario",
                "operationId": "27f883ec77f55250d30be38923ba5bf3",
                "responses": {
                    "200": {
                        "description": "Comentario eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}